Moved product-tier-type.enum.ts to billing folder
