+++ apps/web/src/app/auth/trial-initiation/secrets-manager/secrets-manager-trial-paid-stepper.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, ViewChild } from "@angular/core";
 
-import { ProductTierType } from "@bitwarden/common/enums";
+import { ProductTierType } from "@bitwarden/common/billing/enums";
 
 import {
   OrganizationCreatedEvent,
