+++ apps/web/src/app/auth/trial-initiation/trial-initiation.component.ts
@@ -9,8 +9,7 @@ import { PolicyApiServiceAbstraction } from "@bitwarden/common/admin-console/abs
 import { PolicyService } from "@bitwarden/common/admin-console/abstractions/policy/policy.service.abstraction";
 import { MasterPasswordPolicyOptions } from "@bitwarden/common/admin-console/models/domain/master-password-policy-options";
 import { Policy } from "@bitwarden/common/admin-console/models/domain/policy";
-import { PlanType } from "@bitwarden/common/billing/enums";
-import { ProductTierType } from "@bitwarden/common/enums";
+import { PlanType, ProductTierType } from "@bitwarden/common/billing/enums";
 import { ReferenceEventRequest } from "@bitwarden/common/models/request/reference-event.request";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
 import { LogService } from "@bitwarden/common/platform/abstractions/log.service";
@@ -67,7 +66,7 @@ export class TrialInitiationComponent implements OnInit, OnDestroy {
   billingSubLabel = "";
   layout = "default";
   plan: PlanType;
-  product: ProductTierType;
+  productTier: ProductTierType;
   accountCreateOnly = true;
   useTrialStepper = false;
   policies: Policy[];
@@ -153,16 +152,16 @@ export class TrialInitiationComponent implements OnInit, OnDestroy {
 
         if (this.org === ValidOrgParams.families) {
           this.plan = PlanType.FamiliesAnnually;
-          this.product = ProductTierType.Families;
+          this.productTier = ProductTierType.Families;
         } else if (this.org === ValidOrgParams.teamsStarter) {
           this.plan = PlanType.TeamsStarter;
-          this.product = ProductTierType.TeamsStarter;
+          this.productTier = ProductTierType.TeamsStarter;
         } else if (this.org === ValidOrgParams.teams) {
           this.plan = PlanType.TeamsAnnually;
-          this.product = ProductTierType.Teams;
+          this.productTier = ProductTierType.Teams;
         } else if (this.org === ValidOrgParams.enterprise) {
           this.plan = PlanType.EnterpriseAnnually;
-          this.product = ProductTierType.Enterprise;
+          this.productTier = ProductTierType.Enterprise;
         }
       } else if (this.routeFlowOrgs.includes(qParams.org)) {
         this.referenceData.flow = qParams.org;
@@ -268,11 +267,11 @@ export class TrialInitiationComponent implements OnInit, OnDestroy {
   }
 
   get trialOrganizationType(): TrialOrganizationType {
-    switch (this.product) {
+    switch (this.productTier) {
       case ProductTierType.Free:
         return null;
       default:
-        return this.product;
+        return this.productTier;
     }
   }
