+++ libs/common/src/admin-console/models/domain/organization.ts
@@ -1,6 +1,6 @@
 import { Jsonify } from "type-fest";
 
-import { ProductTierType } from "../../../enums";
+import { ProductTierType } from "../../../billing/enums";
 import { OrganizationUserStatusType, OrganizationUserType, ProviderType } from "../../enums";
 import { PermissionsApi } from "../api/permissions.api";
 import { OrganizationData } from "../data/organization.data";
