+++ libs/common/src/admin-console/models/response/profile-organization.response.ts
@@ -1,4 +1,4 @@
-import { ProductTierType } from "../../../enums";
+import { ProductTierType } from "../../../billing/enums";
 import { BaseResponse } from "../../../models/response/base.response";
 import { OrganizationUserStatusType, OrganizationUserType, ProviderType } from "../../enums";
 import { PermissionsApi } from "../api/permissions.api";
