+++ libs/common/src/billing/enums/index.ts
@@ -3,3 +3,4 @@ export * from "./plan-sponsorship-type.enum";
 export * from "./plan-type.enum";
 export * from "./transaction-type.enum";
 export * from "./bitwarden-product-type.enum";
+export * from "./product-tier-type.enum";
