+++ libs/common/src/enums/index.ts
@@ -6,4 +6,3 @@ export * from "./http-status-code.enum";
 export * from "./integration-type.enum";
 export * from "./native-messaging-version.enum";
 export * from "./notification-type.enum";
-export * from "./product-tier-type.enum";
